Version 0.0.18
Refactor & UI fixes
